requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pyjwt